Split standard library bypass evidence
