+++ test/reach/evidence/standard_library_bypass/enum_test.exs
@@ -1,0 +1,42 @@
+defmodule Reach.Evidence.StandardLibraryBypass.EnumTest do
+  use ExUnit.Case, async: true
+
+  alias Reach.Evidence.StandardLibraryBypass.Enum, as: EnumEvidence
+
+  defp collect(ast) do
+    {_ast, evidence} =
+      Macro.prewalk(ast, [], fn node, acc -> {node, EnumEvidence.collect_node(node, acc)} end)
+
+    Enum.reverse(evidence)
+  end
+
+  test "collects direct and reduce-based flat_map evidence" do
+    ast = Code.string_to_quoted!("items |> Enum.map(&expand/1) |> List.flatten()")
+    assert [%{kind: :manual_flat_map}] = collect(ast)
+
+    ast =
+      Code.string_to_quoted!("Enum.reduce(items, [], fn item, acc -> acc ++ expand(item) end)")
+
+    assert [%{kind: :manual_flat_map_reduce}] = collect(ast)
+  end
+
+  test "collects order-safe prepend reverse flat_map evidence" do
+    ast =
+      Code.string_to_quoted!("""
+      items
+      |> Enum.reduce([], fn item, acc -> Enum.reverse(expand(item), acc) end)
+      |> Enum.reverse()
+      """)
+
+    assert [%{kind: :manual_flat_map_prepend_reverse}] = collect(ast)
+  end
+
+  test "collects frequencies evidence" do
+    ast =
+      Code.string_to_quoted!("""
+      Enum.reduce(items, %{}, fn item, acc -> Map.update(acc, item, 1, &(&1 + 1)) end)
+      """)
+
+    assert [%{kind: :manual_frequencies}] = collect(ast)
+  end
+end
